+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### Fixed
 - Made APC store setMulti with integer keys
 - Short-circuit getting/setting/deleting multiple keys
+- Make sure SQLite treats numerical keys as strings, not numbers
 
 
 ## [1.4.2] - 2017-02-07