+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Short-circuit getting/setting/deleting multiple keys
 - Make sure SQLite treats numerical keys as strings, not numbers
 - Support numerical keys with Couchbase
+- Fix how Couchbase (failed to) handle negative TTLs
 
 
 ## [1.4.2] - 2017-02-07